--- apps/web/src/app/tournament/[id]/create/_types/tournament.ts
@@ -1,3 +1,0 @@
-export type PostTournamentItemLinkResponseT = {
-  tournamentItemId: number;
-};
