+++ apps/web/src/hooks/usePostTournamentItemLink.ts
@@ -3,12 +3,11 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useRouter } from 'next/navigation';
 import { toast } from 'sonner';
 
+import { postTournamentItemLink } from '@/apis/postTournamentItemLink';
 import { QUERY_ACTION } from '@/consts/queryAction';
 import { ROUTES } from '@/consts/route';
 import { getApiErrorCode, getApiErrorStatus, isGlobalNetError } from '@/utils/apiError';
 import { getApiErrorMessage } from '@/utils/getApiErrorMessage';
-
-import { postTournamentItemLink } from '../_apis/postTournamentItemLink';
 
 type UsePostTournamentItemLinkOptionsT = {
   /** 입력 폼처럼 에러를 화면 안에서 안내하는 경우 false — 4xx 토스트를 끈다 */
